+++ includes/frontend/post-options.php
@@ -34,7 +34,6 @@ function lmt_print_last_modified_info_post( $content ) {
     }
 
     if( isset( $modified_content ) ) {
-
         if( isset($options['lmt_show_last_modified_time_date_post']) && ($options['lmt_show_last_modified_time_date_post'] == 'before_content') ) {
             $fullcontent = $modified_content . $content; 
         }
@@ -44,13 +43,22 @@ function lmt_print_last_modified_info_post( $content ) {
     } 
 
     if ( isset( $fullcontent ) && is_singular( 'post' ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
+        if( apply_filters( 'wplmi_disable_post_output', false ) ) {
+            return $content;
+        }
         return $fullcontent;
+    }
+
+    if( isset($options['lmt_show_on_homepage']) && ($options['lmt_show_on_homepage'] == 'yes') ) {
+        if ( isset( $fullcontent ) && ( is_home() || is_author() || is_category() || is_tag() ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
+            return $fullcontent;
+        }
     }
 
     if( isset($options['lmt_custom_post_types_list']) ) {
         $post_types = $options['lmt_custom_post_types_list'];
-        foreach($post_types as $item) {
-            if ( isset( $fullcontent ) && is_singular( $item ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
+        foreach( $post_types as $post_type ) {
+            if ( isset( $fullcontent ) && is_singular( $post_type ) && !get_post_meta( get_the_ID(), '_lmt_disable', true ) == 'yes' ) { 
                 return $fullcontent;
             }
         }